+++ src/plugins/nonebot_bison/platform/bilibili.py
@@ -5,7 +5,7 @@ from typing import Any, Optional
 from ..post import Post
 from ..types import Category, RawPost, Tag, Target
 from ..utils import http_client
-from .platform import CategoryNotSupport, NewMessage
+from .platform import CategoryNotSupport, NewMessage, StatusChange
 
 
 class Bilibili(NewMessage):
@@ -155,3 +155,73 @@ class Bilibili(NewMessage):
         else:
             raise CategoryNotSupport(post_type)
         return Post("bilibili", text=text, url=url, pics=pic, target_name=target_name)
+
+
+class Bilibililive(StatusChange):
+    # Author : Sichongzou
+    # Date : 2022-5-18 8:54
+    # Description : bilibili开播提醒
+    # E-mail : 1557157806@qq.com
+    categories = {}
+    platform_name = "bilibili-live"
+    enable_tag = False
+    enabled = True
+    is_common = True
+    schedule_type = "interval"
+    schedule_kw = {"seconds": 10}
+    name = "Bilibili直播"
+    has_target = True
+
+    async def get_target_name(self, target: Target) -> Optional[str]:
+        async with http_client() as client:
+            res = await client.get(
+                "https://api.bilibili.com/x/space/acc/info", params={"mid": target}
+            )
+            res_data = json.loads(res.text)
+            if res_data["code"]:
+                return None
+            return res_data["data"]["name"]
+
+    async def get_status(self, target: Target):
+        async with http_client() as client:
+            params = {"mid": target}
+            res = await client.get(
+                "https://api.bilibili.com/x/space/acc/info",
+                params=params,
+                timeout=4.0,
+            )
+            res_dict = json.loads(res.text)
+            if res_dict["code"] == 0:
+                info = {}
+                info["uid"] = res_dict["data"]["mid"]
+                info["uname"] = res_dict["data"]["name"]
+                info["live_state"] = res_dict["data"]["live_room"]["liveStatus"]
+                info["room_id"] = res_dict["data"]["live_room"]["roomid"]
+                info["title"] = res_dict["data"]["live_room"]["title"]
+                info["cover"] = res_dict["data"]["live_room"]["cover"]
+                return info
+            else:
+                return []
+
+    def compare_status(self, target: Target, old_status, new_status) -> list[RawPost]:
+        if (
+            new_status["live_state"] != old_status["live_state"]
+            and new_status["live_state"] == 1
+        ):
+            return [new_status]
+        else:
+            return []
+
+    async def parse(self, raw_post: RawPost) -> Post:
+        url = "https://live.bilibili.com/{}".format(raw_post["room_id"])
+        pic = [raw_post["cover"]]
+        target_name = raw_post["uname"]
+        title = raw_post["title"]
+        return Post(
+            self.name,
+            text=title,
+            url=url,
+            pics=pic,
+            target_name=target_name,
+            compress=True,
+        )